pagination and refreshing working

+++ app/src/main/java/com/example/bit603_a3_tonylawrie/ShowInventory.java
@@ -16,12 +16,12 @@ import com.google.android.material.bottomnavigation.BottomNavigationView;
 import java.util.List;
 
 public class ShowInventory extends AppCompatActivity {
-  private static String TAG = "InventoryFragment";
+  private static final String TAG = "Inventory";
 
   BottomNavigationView inventoryNavigationView;
   Menu inventoryMenu;
-  MenuItem previous,next;
-  TextView listText;
+  MenuItem previous, next;
+  TextView listText, pageCount;
   private int pageSize, page, lastpage;
   List<Inventory> inventory;
 
@@ -40,65 +40,60 @@ public class ShowInventory extends AppCompatActivity {
     listText = findViewById(R.id.inventory_list);
     previous = inventoryMenu.findItem(R.id.previous);
     next = inventoryMenu.findItem(R.id.next);
+    pageCount = findViewById(R.id.page_counter);
 
     pageSize = 5;
     page = 1;
-
+    lastpage = inventory.size() / pageSize;
 
     //initial call to update menu to set the disable options
     update();
     //item selected listener
     inventoryNavigationView.setOnItemSelectedListener(item -> {
-      switch (item.getItemId()) {
-        case R.id.next:
-          page++;
-          setItems();
-          break;
-        case R.id.add_item:
-          break;
-        case R.id.add_test_data:
-          addSampleData();
-          break;
-        case R.id.delete_items:
-          deleteData();
-          break;
-        case R.id.previous:
-          page--;
-          setItems();
-          break;
+      if(item.getItemId() == R.id.next) {
+        page++;
+        setItems();
+      } else if(item.getItemId() == R.id.add_item){
+        //TODO
+      } else if(item.getItemId() == R.id.add_test_data){
+        addSampleData();
+      } else if(item.getItemId() == R.id.delete_items){
+        deleteData();
+      } else if(item.getItemId() == R.id.previous){
+        page--;
+        setItems();
       }
       return true;
     });
 
-    //  if inventory is empty display message otherwise get 1st set of data
-    if (inventory.size() <= 0) {
-      listText.setText(R.string.empty_inventory);
-    } else {
-      // call getItems() to get initial set of items
-      setItems();
-    }
+    // call getItems() to get initial set of items
+    setItems();
   }
 
   // set 5 items at a time
   private void setItems() {
-    Log.d(TAG, "setItems: called");
-    listText.setText("");
-    String item, type, quantity, outString;
+    if (inventory.size() <= 0) {
+      listText.setText(R.string.empty_inventory);
+    } else {
+      Log.d(TAG, "setItems: called");
+      listText.setText("");
+      String item, type, quantity, outString;
 
 //    Log.d(TAG, "I: " + String.valueOf((page - 1) * pageSize));
 //    Log.d(TAG, "page" + page);
 //    Log.d(TAG, "pagesize: " + pageSize);
 //    Log.d(TAG, "condition: " + page * pageSize);
 
-    for (int i = (page - 1) * pageSize; i < page * pageSize - 1; i++) {
-      if (i < inventory.size() && i >= 0) {
+      for (int i = (page - 1) * pageSize; i < page * pageSize - 1; i++) {
+        if (i < inventory.size() && i >= 0) {
 //        Log.d(TAG, String.valueOf(i));
-        item = inventory.get(i).getItem();
-        type = inventory.get(i).getType();
-        quantity = String.valueOf(inventory.get(i).getQuantity());
+          item = inventory.get(i).getItem();
+          type = inventory.get(i).getType();
+          quantity = String.valueOf(inventory.get(i).getQuantity());
 
-        outString = item + "\u0009" + type + "\u0009Quantity: " + quantity + "\r\n";
-        listText.append(outString);
+          outString = item + "\u0009" + type + "\u0009Quantity: " + quantity + "\r\n";
+          listText.append(outString);
+        }
       }
     }
     update();
@@ -116,13 +111,13 @@ public class ShowInventory extends AppCompatActivity {
       Toast toast = Toast.makeText(this, "Inventory deleted!", Toast.LENGTH_SHORT);
       toast.setGravity(Gravity.TOP | Gravity.CENTER_HORIZONTAL, 0, 40);
       toast.show();
-      setItems();
+      refreshActivity();
+
     });
     // negative confirmation
     builder.setNegativeButton("Cancel", null);
     AlertDialog confirmation = builder.create();
     confirmation.show();
-
   }
 
 
@@ -140,29 +135,30 @@ public class ShowInventory extends AppCompatActivity {
       Toast toast = Toast.makeText(this, "Sample Data added successfully!", Toast.LENGTH_SHORT);
       toast.setGravity(Gravity.TOP | Gravity.CENTER_HORIZONTAL, 0, 40);
       toast.show();
-      setItems();
+      refreshActivity();
     });
     // negative confirmation
     builder.setNegativeButton("Cancel", null);
-
     AlertDialog confirmation = builder.create();
     confirmation.show();
-
   }
 
   private void update() {
-    lastpage = inventory.size() / pageSize;
-
-    if (page <= 1) {
-      previous.setEnabled(false);
-    } else {
-      previous.setEnabled(true);
+    String pageText = page + "/" + lastpage;
+    pageCount.setText(pageText);
+    if (inventory.size() <= 0) {
+      listText.setText(R.string.empty_inventory);
     }
+    // if on first page set disable previous
+    previous.setEnabled(page > 1);
+    // if on last page disable next
+    next.setEnabled(page < lastpage);
+  }
 
-    if (page >= lastpage) {
-      next.setEnabled(false);
-    } else {
-      next.setEnabled(true);
-    }
+  private void refreshActivity() {
+    finish();
+    overridePendingTransition(0, 0);
+    startActivity(getIntent());
+    overridePendingTransition(0, 0);
   }
 }
